requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken
psycopg2-binary
python-dotenv
